[core] Split the monolithic Supervisor into a SupervisorCore runtime and facade
